requirements.txt
asgiref==3.6.0
autopep8==2.0.1
coverage==7.1.0
Django==4.1.5
djangorestframework==3.14.0
html-json-forms==1.1.1
natural-keys==2.0.0
pycodestyle==2.10.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1